+++ data/scripts/libs/lib3.js
@@ -1,0 +1,3 @@
+const three = 3;
+
+export default three;