+++ data/scripts/libs/libTwo.js
@@ -1,3 +1,0 @@
-const dummy = 4;
-
-export default dummy;