+++ data/scripts/main-backoffice.js
@@ -1,4 +1,8 @@
-require.ensure(['./libs/libTwo'], () => {
-	const su = require('./libs/libTwo'); // eslint-disable-line global-require
+require.ensure(['./libs/lib4'], () => {
+	const su = require('./libs/lib4'); // eslint-disable-line global-require
+	console.log(su.default);
+});
+require.ensure(['./libs/lib5'], () => {
+	const su = require('./libs/lib5'); // eslint-disable-line global-require
 	console.log(su.default);
 });