+++ data/scripts/webpack.dev.config.js
@@ -2,6 +2,7 @@ const webpackConfig = require('./webpack-config');
 
 module.exports = webpackConfig({
 	hot: false,
+	// Don’t use [chunkhash] in development since this will increase compilation time
 	hash: false,
 	debug: true,
 	optimize: false,