+++ src/build.js
@@ -1,27 +1,26 @@
-/**
-*
-* Copyright © 2014-2016 killanaca All rights reserved.
-*
-* This source code is licensed under the MIT license found in the
-* LICENSE.txt file in the root directory of this source tree.
-*
-*/
-
 import c from './libs/config';
 import clean from './clean';
 import buildJS from './buildJs';
-import imagemin from './imagemin';
-import buildLess from './buildLess';
+import buildCss from './buildCss';
+import buildImages from './buildImages';
 
+
+/**
+ * Build js, css, images.
+ * This task could be called directly
+ * @return {Promise} A Promise
+ * @example <caption>run this on your terminal</caption>
+ * node src/run build
+ */
 async function build() {
 	const config = c().load();
 	// clean folder
-	await clean(config);
-	// compile js files
-	await Promise.all([
-		buildJS(config),
-		buildLess(config),
-		imagemin(config),
+	await clean({ config });
+	// compile css, js, image files in parallel
+	return Promise.all([
+		buildJS({ config, cleaned: true }),
+		buildCss({ config, cleaned: true }),
+		buildImages({ config, cleaned: true }),
 	]);
 }