+++ src/bump.js
@@ -1,14 +1,14 @@
-/**
-*
-* Copyright © 2014-2016 killanaca All rights reserved.
-*
-* This source code is licensed under the MIT license found in the
-* LICENSE.txt file in the root directory of this source tree.
-*/
 import fu from './libs/files-utils';
 import pkg from '../package.json';
 import c from './libs/config';
 
+/**
+ * Upgrade version inside package.json and web.config ([OPT-IN]) according semver.
+ * This task could be called directly
+ * @return {Promise} A Promise
+ * @example <caption>run this on your terminal</caption>
+ * node src/run bump
+ */
 async function bump() {
 	const config = c().load();
 	const versionRequest = process.argv[process.argv.length - 1];