+++ src/clean.js
@@ -1,12 +1,3 @@
-/**
-*
-* Copyright © 2014-2016 killanaca All rights reserved.
-*
-* This source code is licensed under the MIT license found in the
-* LICENSE.txt file in the root directory of this source tree.
-*
-*/
-
 import del from 'del';
 import fs from './libs/fs';
 import trimEnd from 'lodash/trimEnd';
@@ -16,7 +7,14 @@ const debug = require('debug')('dt');
 
 const loadConfig = () => c().load();
 
-async function clean(config = loadConfig()) {
+/**
+ * Delete and recreate folders
+ * This task could be called directly
+ * @param {object} [obj] - obj
+ * @param {object} obj.config - The config Object
+ * @return {Promise} A Promise
+ */
+async function clean({ config = loadConfig() } = {}) {
 	const distFolder = [];
 	// delete js build folder?
 	if (config.get('buildPathJs')) {
@@ -24,7 +22,7 @@ async function clean(config = loadConfig()) {
 	}
 
 	// delete css build folder?
-	if (!config.get('preserveBuildPathCss')) {
+	if (!config.get('preserveBuildPathCss') && config.get('buildPathCss')) {
 		distFolder.push(path.join(process.cwd(), config.get('buildPathCss')));
 	}