+++ src/libs/fs.js
@@ -1,11 +1,3 @@
-/**
-*
-* Copyright © 2014-2016 killanaca All rights reserved.
-*
-* This source code is licensed under the MIT license found in the
-* LICENSE.txt file in the root directory of this source tree.
-*/
-
 import fs from 'fs';
 import mkdirp from 'mkdirp';
 import isString from 'lodash/isString';
@@ -14,10 +6,7 @@ import { promisify } from 'bluebird';
 import path from 'path';
 const debug = require('debug')('dt');
 
-
 // NOTE: Promise.promisify takes a function that takes a callback as its last argument and converts it into a function that returns a promise (without the need for a callback)
-
-
 const makeDirAsync = promisify(mkdirp);
 const fileExistsPromise = promisify(fs.access);
 
@@ -33,6 +22,13 @@ async function writeFileAsync(filepath, content) {
 	const fullPath = path.join(process.cwd(), filepath);
 	debug(`try to write a file at path ${fullPath}`);
 	return await writeFilep(fullPath, content);
+}
+
+function writeAsync(filepath, content) {
+	const writeFilep = promisify(fs.writeFile);
+	const fullPath = path.join(process.cwd(), filepath);
+	debug(`try to write a file at path ${fullPath}`);
+	return writeFilep(fullPath, content);
 }
 
 async function readFileAsync(filepath) {
@@ -68,17 +64,6 @@ async function fileExistsAsync(filepath) {
 }
 
 /**
- * Check if file exists on the specified path
- * @param  {String} path The filepath
- * @return {Promise}      A promise that returns true if the file exists, false otherwise
- */
-function fileExistsPr(path) {
-	return fileExistsPromise(path, fs.R_OK)
-	.then(() => true)
-	.catch(Error, () => false);
-}
-
-/**
 * Create folder(s) from a string path or an array of string path
 * @param  {String|Array} paths a string path or an array of string path
 * @return {Promise}      The promise object that is resolved when all the files are created
@@ -93,6 +78,7 @@ const makeDirsAsync = (paths) => {
 };
 
 export default {
+	writeAsync,
 	writeFileAsync,
 	writeFileSync: fs.writeFileSync,
 	readFileAsync,
@@ -100,5 +86,4 @@ export default {
 	readFileSync: fs.readFileSync,
 	makeDirsAsync,
 	fileExistsAsync,
-	fileExistsPr,
 };