+++ src/watch.js
@@ -1,32 +1,46 @@
-/**
-*
-* Copyright © 2014-2016 killanaca All rights reserved.
-*
-* This source code is licensed under the MIT license found in the
-* LICENSE.txt file in the root directory of this source tree.
-*
-*/
-
 import c from './libs/config';
 import webpackDevServer from './utils/webpack-dev-server';
 import lessTask from './utils/less';
+import sassTask from './utils/sass';
 import chokidar from 'chokidar';
 import logger from './libs/logger';
 import path from 'path';
 
 const loadConfig = () => c().load();
 
-async function watch(config = loadConfig()) {
+/**
+ * Watch JavaScript (via webpack), less/sass folder.
+ * This task could be called directly
+ * @param {object} [obj] - obj
+ * @param {object} obj.config - The config Object
+ * @return {Promise} A Promise
+ */
+async function watch({ config = loadConfig() } = {}) {
 	const tasks = [];
-	// compile less the first time
-	tasks.push(lessTask(config));
-	const watchLess = path.join(process.cwd(), config.get('srcLess'));
-	chokidar.watch(watchLess).on('change', (filepath) => {
-		logger.log(`${filepath} changed`);
-		lessTask(config);
-	});
 	// add webpack to task. watch and compile js files
 	tasks.push(webpackDevServer(config));
+
+	// watch less files???
+	if (config.get('srcLess')) {
+		// compile less files for the first time
+		tasks.push(lessTask(config, false));
+		const watchFiles = path.join(process.cwd(), config.get('srcLess'));
+		chokidar.watch(watchFiles).on('change', (filepath) => {
+			logger.log(`${filepath} less file changed`);
+			lessTask(config);
+		});
+	}
+	// watch sass files???
+	if (config.get('srcSass')) {
+		// compile sass files for the first time
+		tasks.push(sassTask(config, false));
+		const watchFiles = path.join(process.cwd(), config.get('srcSass'));
+		chokidar.watch(watchFiles).on('change', (filepath) => {
+			logger.log(`${filepath} sass file changed`);
+			sassTask(config);
+		});
+	}
+
 	await Promise.all(tasks);
 }