+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+<a name="1.2.0"></a>
+# [1.2.0](https://github.com/killanaca/deployment-tools/compare/v1.1.0...v1.2.0) (2016-06-28)
+
+
+### Bug Fixes
+
+* **c:** fixed an issue with image url inside less files regarding to cdn ([f5cef3d](https://github.com/killanaca/deployment-tools/commit/f5cef3d))
+* **css:** adjust image urls inside css for CDN ([2ad3bb6](https://github.com/killanaca/deployment-tools/commit/2ad3bb6))
+* **image:** process svg image ([92ff42b](https://github.com/killanaca/deployment-tools/commit/92ff42b))
+* **less:** enable minify on build not in watch mode ([782da5a](https://github.com/killanaca/deployment-tools/commit/782da5a))
+* **less:** fixed an issue with image relative path ([8933ba1](https://github.com/killanaca/deployment-tools/commit/8933ba1))
+* **less:** fixed an issue with image url inside less ([15484d5](https://github.com/killanaca/deployment-tools/commit/15484d5))
+* **less:** preserve css folder (no clean) if preserveBuildPathCss equals true ([ac47649](https://github.com/killanaca/deployment-tools/commit/ac47649))
+* **lib:** added missing modernizr-webpack-plugin ([118de29](https://github.com/killanaca/deployment-tools/commit/118de29))
+* **upload:** fixed an issue with the path of files on Azure Storage ([0c9ae2b](https://github.com/killanaca/deployment-tools/commit/0c9ae2b))
+* **webpackserv:** fixed an issue with webpack in dev mode ([d20df8f](https://github.com/killanaca/deployment-tools/commit/d20df8f))
+
+
+### Features
+
+* **core:** added imagemin task ([04fff98](https://github.com/killanaca/deployment-tools/commit/04fff98))
+* **core:** added lint features to css task, some refactoring ([56185bb](https://github.com/killanaca/deployment-tools/commit/56185bb))
+* **core:** first commit ([a5612b8](https://github.com/killanaca/deployment-tools/commit/a5612b8))
+* **js:** added long term caching ([995618c](https://github.com/killanaca/deployment-tools/commit/995618c))
+* **postinstall:** add postinstall code ([33c1b3a](https://github.com/killanaca/deployment-tools/commit/33c1b3a))
+
+
+
 <a name="1.0.0"></a>
 # 1.0.0 (2016-06-06)